+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Flow Connector for Magento 2 Change Log
 
+## 2.5.0
+- Add Sync Stream registration to intialization command
+- Add Sync Stream Record registration for order import events
+
 ## 2.4.2
 - Add tracking numbers and URLs directly to orders on label webhook processing
 - Clear carts via Continue Shopping redirect